small refactoring and added check for open queries of the test-oracle
