+++ src/main/java/com/kryptokrauts/aeternity/sdk/service/aeternal/AeternalService.java
@@ -1,7 +1,7 @@
 package com.kryptokrauts.aeternity.sdk.service.aeternal;
 
 import com.kryptokrauts.aeternity.sdk.service.aeternal.domain.ActiveAuctionsResult;
-import com.kryptokrauts.aeternity.sdk.service.aeternal.order.NameSortBy;
+import com.kryptokrauts.aeternity.sdk.service.aeternal.domain.NameSortBy;
 import java.io.IOException;
 import java.math.BigInteger;
 import java.util.Optional;
