+++ src/main/java/com/kryptokrauts/aeternity/sdk/service/aeternity/impl/AeternityService.java
@@ -4,8 +4,6 @@ import com.kryptokrauts.aeternity.generated.api.ExternalApiImpl;
 import com.kryptokrauts.aeternity.generated.api.rxjava.ExternalApi;
 import com.kryptokrauts.aeternity.sdk.service.account.AccountService;
 import com.kryptokrauts.aeternity.sdk.service.account.impl.AccountServiceImpl;
-import com.kryptokrauts.aeternity.sdk.service.aens.NameService;
-import com.kryptokrauts.aeternity.sdk.service.aens.impl.NameServiceImpl;
 import com.kryptokrauts.aeternity.sdk.service.aeternal.AeternalService;
 import com.kryptokrauts.aeternity.sdk.service.aeternal.impl.AeternalServiceImpl;
 import com.kryptokrauts.aeternity.sdk.service.aeternity.AeternityServiceConfiguration;
@@ -13,6 +11,8 @@ import com.kryptokrauts.aeternity.sdk.service.compiler.CompilerService;
 import com.kryptokrauts.aeternity.sdk.service.compiler.impl.SophiaCompilerServiceImpl;
 import com.kryptokrauts.aeternity.sdk.service.info.InfoService;
 import com.kryptokrauts.aeternity.sdk.service.info.impl.InfoServiceImpl;
+import com.kryptokrauts.aeternity.sdk.service.name.NameService;
+import com.kryptokrauts.aeternity.sdk.service.name.impl.NameServiceImpl;
 import com.kryptokrauts.aeternity.sdk.service.oracle.OracleService;
 import com.kryptokrauts.aeternity.sdk.service.oracle.impl.OracleServiceImpl;
 import com.kryptokrauts.aeternity.sdk.service.transaction.TransactionService;
